RocksDB has hint policy for fcntl to decrease WAF. The native hint policy is WAF is short, L0-L1 is medium, L2 is long and >=L3 is extreme. However, based on our experiments on the FDP SSD, the reduction in the WAF is not significant.
As we observed the data lifetime of each levels, we changed the hint policy for fcntl. The new policy is WAF is short, L0-L3 is medium, L4 is long and >=L5 is extreme. The WAF gained reduction dramatically with experiments.

| **Lifetime** | **Hint  Policy (native, num_levels <=5)** | **Hint Policy (optimized, num_levels >=6)** |
| ------------ | ----------------------------------------- | ------------------------------------------- |
| WLTH_SHORT   | WAL                                       | WAL                                         |
| WLTH_MEDIUM  | Level0  ~  Level1                         | Level0  ~  Level3                           |
| WLTH_LONG    | Level2                                    | Level4                                      |
| WLTH_EXTREME | >=  Level3                                | >=  Level5                                  |

e.g. With YCSB workload, load 200 million records and update 100 million records, RocksDB WAF on ext4 was reduced from 2.7 (legacy SSD) to 2.4 (FDP SSD) on native hint policy, but from 2.7 (legacy SSD) to 1.9 (FDP SSD).

The patch differentiates cases based on the num_levels option configuration. If the num_levels option is configured <=5, keeps the native hint policy. If the num_levels option is configured >=6, use the new hint policy.

 The proportion of trivial move is small and it would not affect the WAF result.