Add GetOptionConfigName() helper for test debugging
Summary:

Challenge: Many tests in RocksDB test different DB option configs in a single test run using a for loop. When a test that cycles through different configs fails, it's difficult to know which config failed as there's no obvious way to see this in the stack trace.

Improvement: Added GetOptionConfigName() to convert an option config int to string (readable name). This method can be used to log the config names being tested, making it easy to identify at which point the test fails.