Fix ODR violation from open source folly build, update
Summary: Updating folly to
https://github.com/facebook/folly/commit/8a9fc1e80a18cafadbec85e33d5042ce13a7c634
or beyond was failing an F14Table assertion for a very subtle reason:
ODR violation between the folly build and RocksDB build because folly
build was release mode and RocksDB build was debug mode. What was
happening was that folly change introduced a dependence on kDebug
(whether build is debug) in a hashing implementation in a .h file, and
the inconsistency between the inlined implementation during RocksDB
build and the linked-to implementation from the folly build was leading
to inconsistencies in the data structure.

The fix is to ensure we build folly in debug mode for debug mode RocksDB
builds. I've also updated folly to latest commit here.

Intended follow-up:
* Add release mode folly+RocksDB to CI (might need fixing; haven't
  checked)
* Make the folly build caching more friendly+accurate by hashing the relevant
  Makefile parts and tagging whether debug or release. Not in this PR because
  then you wouldn't be able to see what changed in the folly build steps
  themselves.

Test Plan: manual + CI