Improve distinct compression for index and data blocks
Summary: This change enables a custom CompressionManager / Compressor to
adopt custom handling for data and index blocks. In particular, index
blocks for format_version >= 4 use a distinct variant of the block
format. Thus, a potentially format-aware compression algorithm such as
OpenZL should be told which kind of block we are compressing. (And
previously I avoided passing block type in CompressBlock for efficient
handling of things like dictionaries but also avoiding checks on every
CompressBlock call.)

Most of the change is in BlockBasedTableBuilder to call
MaybeCloneSpecialized for both kDataBlock and for kIndexBlock. But I
also needed some small tweaks/additions to the public API also:
* Require a Clone() function from Compressors, to support proper
  implementations of MaybeCloneSpecialized() in wrapper Compressors.
* Assert that the default implementation of
  CompressorWrapper::MaybeCloneSpecialized() is only used in allowable
  cases.
* Convenience function Compressor::CloneMaybeSpecialized()

Test Plan: Greatly expanded DBCompressionTest.CompressionManagerWrapper
to make sure the distinction between data blocks and index blocks is
properly communicated to a custom CompressionManager/Compressor. The
test includes processing the expected structure of data and index
blocks, to serve as a tested example for structure-aware compressors.