Relax an assertion related to parallel compression
Summary: Saw a mysterious failure of assertion
`assert(rep_->props.num_data_blocks == 0)` in
DBCompressionTest/CompressionFailuresTest.CompressionFailures/45. This
seems to be caused by a parallel compression failure arriving after the
emit thread has started Finish() but before the Flush() at the start of
Finish(). We can fix this by relaxing the assertion to allow for the
!ok() case. Testing revealed more ok() assertions that needed to be
relaxed/moved.

Test Plan: Added a sync point to inject a failure status in the right
place and added to unit test to be sure the case is essentially covered.
It would arguably be a more realistic test to force a particular thread
interleaving but I believe simple is good here.