Fix AutoSkipCompressorWrapper with new logic
Summary: ... from #14140. The assertion in the default implementation of
CompressorWrapper::MaybeCloneSpecialized() could fail because this
wrapper wasn't overriding it when it should. (See the NOTE on that
implementation.)

Because this release already has a breaking modification to the Compressor API
(adding Clone()), I took this opportunity to add 'const' to
MaybeCloneSpecialized(). Also marked some compression classes as 'final'
that could be marked as such.

Test Plan: unit test expanded to cover this case (verified failing
before). Audited the rest of our CompressorWrappers.